Crud para directivos

+++ front/src/controllers/form_informacionView.js
@@ -4,13 +4,14 @@ import {
   eliminarGobernacionId,
   getConsejoDataId,
   getContraloriaDataId,
+  getDirectivoDataId,
   getGobernacionDataId,
 } from '../api/form_informacion.js'
 import {
   ejerciciosLista,
   validarEjercicioActual,
 } from '../components/form_ejerciciosLista.js'
-import { form_informacionConsejoForm } from '../components/form_informacion/form_informacionConsejoForm.js'
+import { form_informacionDirectivoForm } from '../components/form_informacion/form_informacionDirectivoForm.js'
 import { form_informacionContraloriaForm } from '../components/form_informacion/form_informacionContraloriaForm.js'
 import { form_informacionGobernacionForm } from '../components/form_informacion/form_informacionGobernacionForm.js'
 import { confirmNotification, toastNotification } from '../helpers/helpers.js'
@@ -18,10 +19,12 @@ import { NOTIFICATIONS_TYPES } from '../helpers/types.js'
 import {
   deleteConsejoRow,
   deleteContraloriaRow,
+  deleteDirectivoRow,
   deleteGobernacionRow,
   loadGobernacionTable,
   validateConsejoTable,
   validateContraloriaTable,
+  validateDirectivoTable,
   validateGobernacionTable,
 } from './form_informacionTables.js'
 
@@ -113,7 +116,7 @@ export const validateConsejoView = async () => {
   d.addEventListener('click', async (e) => {
     if (e.target.id === 'consejo-registrar') {
       scroll(0, 0)
-      form_informacionConsejoForm({ elementToInsert: 'consejo-view' })
+      form_informacionDirectivoForm({ elementToInsert: 'consejo-view' })
     }
 
     if (e.target.dataset.editarid) {
@@ -121,7 +124,7 @@ export const validateConsejoView = async () => {
       let data = await getConsejoDataId(e.target.dataset.editarid)
       console.log(data)
 
-      form_informacionConsejoForm({
+      form_informacionDirectivoForm({
         elementToInsert: 'consejo-view',
         data,
       })
@@ -134,6 +137,45 @@ export const validateConsejoView = async () => {
         message: '¿Desea eliminar este registro?',
         successFunction: async function () {
           deleteConsejoRow({ row })
+          eliminarConsejoId(e.target.dataset.eliminarid)
+        },
+      })
+    }
+  })
+}
+
+export const validateDirectivoView = async () => {
+  let btnNewElement = d.getElementById('directivo-registrar')
+
+  validateDirectivoTable()
+
+  console.log('hola')
+
+  d.addEventListener('click', async (e) => {
+    if (e.target.id === 'directivo-registrar') {
+      scroll(0, 0)
+      form_informacionDirectivoForm({ elementToInsert: 'directivo-view' })
+    }
+
+    if (e.target.dataset.editarid) {
+      scroll(0, 0)
+      let data = await getDirectivoDataId(e.target.dataset.editarid)
+      console.log(data)
+
+      form_informacionDirectivoForm({
+        elementToInsert: 'directivo-view',
+        data,
+      })
+    }
+
+    if (e.target.dataset.eliminarid) {
+      let row = e.target.closest('tr')
+
+      confirmNotification({
+        type: NOTIFICATIONS_TYPES.delete,
+        message: '¿Desea eliminar este registro?',
+        successFunction: async function () {
+          deleteDirectivoRow({ row })
           eliminarConsejoId(e.target.dataset.eliminarid)
         },
       })
